WIP: FWLOG: Binary FWaaS logging
Switch FileRuleLogger from text logging with slf4j and Logback to binary logging
using SBE and a BufferedOutputStream. Also add a deserializer that reads the
files. The latter is used only for testing in this patch, but will be used for
the log translation tool in a later patch.

Not yet implemented:

1. Log rotation. This can be done with an external utility like logrotate.

2. Checking header of existing log file. Currently FileRuleLogger will happily
write additional records to any existing file with non-zero size without
checking the header. Given that the log file contains the UUID of the RuleLogger
object, there's no particular reason this file should contain anything other
than what we expect. However, if we ship a new version with a different schema
version, we'll need to check the header to make sure we're not writing log
records with two different schema versions to the same file.

3. Bounds checking: If the total binary-encoded size of the event exceeds 8k,
logging will fail. Since metadata is the only field with arbitrary length and is
not exposed in the API, and the FirewallLogTranslator doesn't create metadata
anywhere close to that size, this shouldn't be a problem, but when we expose
metadata via the API, we'll need to enforce limits.

Change-Id: I9a5d06cd16d34cd30b2298451920f4f1557157d0
Signed-off-by: Brandon Berg <bberg@midokura.com>