10kp: Use scalable state tables
This change replaces the replicated maps in the current state
tables with the scalable state table. This change is mainly done
in the impementation for the four currently existing state tables:

* ArpStateTable
* Ip4MacStateTable
* MacIdStateTable
* MacIp4StateTable

Where the ReplicatedMapStateTable trait is replaced by the
ScalableStateTable trait.

Besides this change, this patch includes the following to
accommodate the transition:

* The state storage getTable() method has changed to allow
  calling the state table constructor with new arguments that
  include the state proxy client and backend connection observable.

* Fixes an issue where the ArpCache would not call start() on the
  underlying table. Since the new state table ignores operations
  unless started, this caused unit tests to fail.

Ref: MI-1003

Change-Id: Ia2c055a06dac04a484e8a0d7026aac5eac042574
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>