FlowState: introduce a manager for file backed ring buffers
Obtaining the metadata for a memory mapped file to create its ring
buffer representation can be a costly operation as it entails scanning
through the whole file reading all headers and picking the first and
last block to update the pointers of the ring buffer.

This patch introduces a flow state manager that keps track of the opened
memory mapped files so they can be reused by several flow state readers
and writers. Concurrency issues between several writers will be dealt in
a later patch.

Change-Id: Ifd144cb5d19b6b213edb20eeccdeb0317b16f56a
Signed-off-by: Xavi León <xavi.leon@midokura.com>