10kp: Scalable state table uses state proxy
With this change the state table adds support for a dual backend,
where the table entries can be read from both the storage as well
as from the state proxy.

The selection of one source over the other is done using the
connection state, where the state proxy takes precedence over the
storage backend. Whenever the state proxy becomes disconnected,
the state table falls-back to the storage.

Ref: MI-1003

Change-Id: Id02e5985f7c0241634170ba997fdbcf564653b9e
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>