Replicated map supports port migration
This fixes an issue in the replicated map, where port migration
was not correctly supported, preventing a host to add an entry
event though the host had no previous ownership of such entry.

To this end, the patch checks when adding and existing entry that
the entry is owned by the host adding it. If that is not the case,
the addition is allowed.

Change-Id: If535160afc4318b4161540b5b6ca508c71fce75a
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>