FlowState: remove curator from minion and mock cassandra
Previously when the flow state minion was deployed on the cluster,
the minion relied on Zookeeper to register the running instance
on the MidoNet discovery service. After moving the minion alongside
the Agent, we don't need Zookeeper anymore as the minion is
accessible through the local interface.

This patch removes the curator framework from the minion and updates the
tests accordingly. It also uses a new (and free) port on each test as it
was the cause of some flakeyness on the CI.

Also, using a real cassandra has been proven to be flakey in the
CI. Mock the cassandra client so we don't rely on a real cassandra
to test the service life cycle of the flow state minion.

Change-Id: I6aea4123a36ddd1f50116554feb97a4e75c7f10d
Signed-off-by: Xavi León <xavi.leon@midokura.com>