Flowstate: remove curator from minion
Previously when the flow state minion was deployed on the cluster,
the minion relied on Zookeeper to register the running instance
on the MidoNet discovery service. After moving the minion alongside
the Agent, we don't need Zookeeper anymore as the minion is
accessible through the local interface.

This patch removes the curator framework from the minion and updates the
tests accordingly. It also uses a new (and free) port on each test as it
was the cause of some flakeyness on the CI.

Change-Id: Ifc8f8def28ee26630236d628ab39601608d38d70
Signed-off-by: Xavi León <xavi.leon@midokura.com>