FlowState: load state locally if there is no previous owner
If a port is bound for the first time to a host, the previous owner
field is not set. Upon reboot, we return an empty flow state batch in
this case instead of checking if we have flow state stored locally from
before the reboot. This patch makes sure we check local storage in this
situation.

Additionally, we start the minions much earlier in the agent
initialization process so when the agent tries to fetch the state from
local storage, the flow state minion is ready to answer. In any case,
increase the number of retries to reduce the likelyness of failure.

Ref.: MI-1190

Change-Id: Ide0b067207d48735142bb3ebb94dd8ed127567d6
Signed-off-by: Xavi León <xavi.leon@midokura.com>