10kp: State tables notify when ready
This change enables a state table to notify an observer when an
initial snapshot was loaded. To this end, the interface adds to
new methods: (i) a ready method that returns an observable that
emits a single notification when the table is ready, and (ii) an
isReady method that returns true if the table is ready.

Fixes: MI-1197

Change-Id: Ia14f4ab698ca0e52b97da2d2769b990a3e8ea5b4
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>