FlowState: flush data in memory upon shutdown
During an scheduled shutdown (not crash), there could be flow state
stored in memory before actually compressing and writing it to the
corresponding memory mapped file. This patch exposes the mapping to
the current active flow state writers so we can flush the in
memory flow state before shutting down the agent.

Ref.: MI-1204 (partially)
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: Ia7f17ac803f36e41ecea44c2d3f6b19b8c6b80a1