FlowState: Don't wait for Cassandra connection
The request proxy was waiting for an established connection with the
cassandra cluster before starting the flow state recovery. This was done
even if the flags to read from cassandra were off. Now if the flags are
off and the Agent is insctructed to go to local storage or to request
flow state from a different Agent then it doesn't wait for cassandra.

Ref: MI-1202

Change-Id: Ica76b34f9a9bc307b8f576c676213bc705343c33
Signed-off-by: Mateo Hermida <mateo@midokura.com>