Agent: wdog kills also agent child process
When wdog kills the Agent, the minions subprocess is not killed as we
just send signals to the Agent process. This patch modifies the wdog
process to notify signals by process group id (which happens to be
the pid of the Agent). This way, the minions subprocess will be able
to execute the shutdown sequence, not being executed otherwise.

This patch also starts the minion subprocess earlies in the boot
sequence so it's ready much earlier to send locally stored flow state.

Ref.: MI-1204 (partially)
Change-Id: I98838a6c4f12c21b56095afdd402122b1aa5b515
Signed-off-by: Xavi León <xavi.leon@midokura.com>