FlowState: read fully the incomming message before resending to minion
Incoming flow state should be sent to the minion to store the messages
locally. However, SBE messages reads the buffer incrementally, updating
the encoded/decoded size as it reads the message. As we need to know the
total size of the message to send it to the minion, just go through the
message until we read it fully.

Change-Id: I7648c16801fb36bae70f2e5f0803bf65963f91fb
Signed-off-by: Xavi León <xavi.leon@midokura.com>