10kp: Fix bridge mapper table notifications
This patch fixes the logic in the bridge mapper notification stream,
where the loading of the bridge's state table was not notified properly.
The current error is that the MAC-port table are initialized in the
final stage of building the bridge device, yet their readiness is
checked in the previous filtering stage.

This patch addresses this, by splitting the final building device
stage in two: (i) one pre-building stage where the bridge's data fields
are computed and the mapper begins loading the MAC-port tables, and (ii)
a final building stage where the mapper creates the immutable
simulation bridge device.

The filtering for the readiness of the state table is therefore moved
between the pre-build and build stages, after the tables begin loading.
Also the readiness observable for the state tables is introduced only
at this stage to prevent recomputation of the bridge data if the only
thing that changes is the readiness of the state tables.

The patch also introduces a state variable to prevent the notification
of duplicate bridge devices, when consecutive notifications come from
the readiness observables and otherwise the bridge has not changed.

Ref: MI-1197

Change-Id: I6c2be31c6bf6f68029142f02419aef2a72c636f1
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit 34f5c9853df2341186a026237995eda9fd679229)