Fix Nova metadata client deserialization
This fixes an issue in a Nova metadata client, where converting
a plaintext response to a string would fail because the appropriate
reader is not configured. The patch reads the data directly from
the input stream and converts it to a string.

Fixes: MI-1207

Change-Id: I3d7d1ebc7c914e0595acf4499718f17a463172a6
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit d9e3241e3a685fbad5dad016ef49f45c359f759e)