FlowState: set the previous owner when creating the bridge port
We don't set the previousHostId field when creating the bridge port,
used by the port mapper to create the simulation port, which is used
to notify the host request proxy of new bindings. This patch makes sure
we set this value so the host request proxy know the previous owner of a
given port to ask for flow state.

Change-Id: If47d057582f04e71281f584a4a73432b441f8657
Signed-off-by: Xavi León <xavi.leon@midokura.com>