FlowState: close files when port is unbound
When ports are unbound from a given host, the flow state files should be
closed (and flushed thereby) so all flow state is written to the memory
mapped files and ready to be read when requested by another minion. This
patch also simplifies the execution flow of the flow state write handler
so it's easier to read.

Change-Id: I932a1336fe0d39373938080c1ff06756fdd0df31
Signed-off-by: Xavi León <xavi.leon@midokura.com>