FlowState: set the previous host id also on binding
From the point of view of midonet, unbinding a port is just clearing the
host field from the port information. Some integration components (e.g.
Nova) does not unbind before binding depending on the operation workflow
(e.g. on port migration). This patch makes sure that we set the
previous host UUID on the port when binding on a host different than the
previous one.

Change-Id: I5906fe09f2163808845e2bcf54751e6899b4a513
Signed-off-by: Xavi León <xavi.leon@midokura.com>