binding: make it python3 compatible
In some newer Linux OS releases, /usr/bin/env python points to python3.
We need to make the binding script compatible or it will just fail to
bind. This change takes care of it by taking advantage of the fact that
the part of httplib that we use is still kept in the new http.client
module in Python3.

Change-Id: Ia71b00ffdfae3e3be5ea15f948a0b780a1dc5a6d
Signed-off-by: Antoni Segura Puimedon <toni@midokura.com>