Log at info level when device not found
Previously, when loading a device that didn't exist, the device
mapper would log an error with a full stack trace. These create
noise when packets for a device that has been deleted are still
in the pipeline.

This patch identifies NotFoundException errors, and logs them at
info level. Also, other errors are downgraded to warnings.

Ref: MI-1099

Change-Id: I9bc0beb800b6dbe31eb9ecc0dc7b1384473dc035
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>