Log at info/debug level when device not found
Previously, when loading a device that didn't exist, the device
mapper would log an error with a full stack trace. These create
noise when packets for a device that has been deleted are still
in the pipeline.

This patch identifies not found errors, and logs them at info
or debug level.

Ref: MI-1099

Change-Id: I099693a3e6770c66a56c2de19b70b8cd8f46c248
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>