Bump base version for compatibility test
There is a patch missing in 1.9.5 (the current version we upgrade from)
that makes postponed simulations to not restart until another event
comes into the back channels in the packet workflow [1]. This makes even
the first check (before testing the upgrade itself) to fail
intermittently. As we have already tested the compatibility of 1.9.5
against the current tip of staging/v1.9, it is safe to bump the base
version to 1.9.6, which already contains the fix.

[1] b842d48eb0b1eca3b8f213f18f989fefdd3810ed

Change-Id: I423a14f67121cf1f353f76130744965affca498f
Signed-off-by: Xavi León <xavi.leon@midokura.com>