Make packet workers use a bounded queue
Previously PacketWorkflow has been an Akka Actor. This means it receives
akka messages for all incoming packages. These akka messages are placed
in a "mailbox", which by default is unbounded in akka. Rather than try
to move to an unbounded queue in akka, this patch modifies the
PacketWorkflow to use a Disruptor RingBuffer for incoming packets. This
ring buffer is bounded. If there are no slots available in the ring
buffer for an incoming packet, the packet is dropped, and a metric is
marked.

This patch also removes a potential deadlock between the PacketWorkflow
thread and FlowProcessor. The FlowProcessor sends messages to
PacketWorkflow over a backchannel, while the PacketWorkflow sends
messages to the FlowProcessor via a Disruptor ring buffer. Previously,
the write to the backchannel was blocking and bounded, so the
PacketWorkflow could block on writing to the flow processor, while the
flow processor is blocked on writing to the backchannel, which would
never unblock as the PacketWorkflow needs to read from it to unblock.

This patch differs from the master patch, in that an addition
back channel was added for the PacketWorkflow, so messages that
previously arrived to the actor, would instead arrive at the back
channel. Also, initialization of the PacketWorkflow is different, as it
has to wait for a message from the DatapathController to process
anything. These interactions are all much cleaner in later branches.

Change-Id: I37f841c13bbd03ced5a54e00e09ca2c546e853fd
Signed-off-by: Ivan Kelly <ivan@midokura.com>