Update patch set 1
Patch Set 1:

The patch feels overengineers in general. We only need one implementation of the metrics class, which takes a MetricsRegistry as parameter. Testing can use these implementations, as we provide a metrics registry for testing. Even better, we can use the metrics in testing to verify things are happening as expected.

In short, remove the MetricRegister, ZoomMetricRegister, and Nil implementations can just call directly to MetricRegistry.

Patch-set: 1
Label: Verified=0