Fix watchdog handling of child processes
This fixes an issue in the watchdog script, where child processes
where managed with an incorrect process group identifier. This
patch ensures that when forking the midolman process, the child
calls the `setsid` method creating a new session and becoming the
leader of a new process group, with the process group taking the
identifier of the child.

Subsequent handling handling of the child processes will use the
new process group identifier.

The process also changes how the watchdog wait for the child
process to terminate, verifying that the process has stopped.

Change-Id: I128b00ca791908755ba56f84ae0b8b9f0c39e8a7
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>