Update patch set 1
Patch Set 1:

it would be good to have an option to the conf tool, so that clients can send us the output. It may be good to output an MD5 hash of the password rather than ****** so that we can verify that the password changes when we change it, and verify what the password is if we have it. For the system property, we should add the option to midolman-env-*.sh so that we don't have to dig into the code to remember how to turn it on.

Patch-set: 1
Reviewer: Ivan Kelly <1003555@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=0