DhcpSubnet does not contain MTU if not set
This changes the DhcpSubnet DTO from the REST API, to not include
the interfaceMTU field, if the field is not set. Previosly, if
the field was not set, the DTO would include it with zero (0) as
default value.

Fixes: MI-1184

Change-Id: I06332f7cd9e64c5f8d197fd9ff5d18f0e26eb824
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit cf04102422f39d8b1da4cc3e5e2694a0068e85a1)