Handle ICMP identifier as a 32 bit integer
Currently the ICMP identifier in a flow key is a 16-bit integer,
which determines some mismatches when comparing the flow key
value with the conntrack value, and also displays ICMP identifiers
higher than 0x7FFF as negative integers.

With this change the ICMP identifier in the flow key becomes a
32 bit integer, and is converted appropriately when read from or
written to the ICMP packet.

Fixes: MI-910

Change-Id: Ia8ce8f39999c57d94bc0b0f2d3b405affeb8faa2
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>