FlowState: protect against reading past the buffer limit
When an agent running on version 1 of the flow state SBE receives a
version 0, it tries to read past the limit of the byte buffer encoding
the flow state message. This makes the actor to crash with an uncaught
exception. This patch makes sure that receiving an older version of a
flow state message is backwards compatible by catching the exception
thrown when reading past the limit, as SBE does not provide any other
mechanism to check for the end of the message.

Change-Id: I0aaf477c5e4cbba951fb3e952266a14098df4e5e
Signed-off-by: Xavi León <xavi.leon@midokura.com>