Fix race condition in ContainerServiceTest
This fixes a race condition in the ContainerService test that
verifies the service remains subscribed to the container scheduler
through a maximum number of failures. The test relies on a service
internal counter to determine that a maximum number of errors
have occurred, and while the maximum has not been reached the
assumption is that the service is still subscribed, and the test
sends an additional error.

However, a previous error that already reached the maximum could
be already scheduled for processing, but hasn't incremented the
error counter yet. Therefore the test assumption may fail at the
current iteration if the service will unsubscribe due to the
previous error before the check.

Change-Id: I30e11e3e2c15ec80c50622648d13ca507c565251
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit 50f72978cde263909d0f1f10ad970fb5bddfa1c5)