FlowState: set timeout when sending requests to minions
Right now, we only set the configured timeout on read operations.
It may be the case that the local minion is not yet ready when
rebooting failing to connect. Although we were wrapping the request in a
retry loop, the retry was made immediately because we didn't set a
timeout on the connect operation. This patch introduces the same timeout
for read and connect operations, giving time to the local minion to
finish the boot sequence.

Change-Id: I65f56cd554cc6d1fd168f15410a3ab2958359f3d
Signed-off-by: Xavi León <xavi.leon@midokura.com>