Fakeagent
Registers an agent in zookeeper, and marks any ports bound to the agent
as up. Useful for giving a simulation a remote endpoint to tunnel
packets to, without having to actually do anything with the packets.
This can be used for benchmarking, etc.

To run, build the midolman super jar, and run the class directly.

./gradlew :midolman:shadowJar
java -cp midolman/build/libs/midolman-<version>-all.jar \
    org.midonet.services.FakeAgent                      \
    -zk <zookeeper> -hostIdFile /tmp/fake-agent.id

Change-Id: I481b8936a157b3d03d9caf348dc7a92097bdaec9
Signed-off-by: Ivan Kelly <ivan@midokura.com>