Fakeagent
Registers an agent in zookeeper, and marks any ports bound to the agent
as up. Useful for giving a simulation a remote endpoint to tunnel
packets to, without having to actually do anything with the packets.
This can be used for benchmarking, etc.

To run, build the midolman super jar, and run the class directly.

./gradlew :midolman:shadowJar
java -cp midolman/build/libs/midolman-<version>-all.jar \
    org.midonet.services.FakeAgent                      \
    -zk <zookeeper> -hostIdFile /tmp/fake-agent.id

Ref MI-1701

Change-Id: I506f2fd2e0e91cc6ee4acead91ebafbc59ba7e4a
Signed-off-by: Ivan Kelly <ivan@midokura.com>