Join the legacy and local flow state requests
We need to serialize the requests for ports so we keep an order on the
flow state we load. The belt is now only waiting on the local storage
future to complete, ignoring the completion of the legacy storage
future. This patch puts those two futures in sequence so the belt won't
move to the next task until both futures have completed.

Change-Id: I3a2c44227dd584776d64ab4312082c90d3aa596f
Signed-off-by: Xavi León <xavi.leon@midokura.com>