Handle exceptions when setting the BGP status
Currently, when setting the BGP status for a connection, the
operation is not protected with an exception handler, causing
any failure to crash the routing handler actor and bring down the
agent.

This patch corrects this, such that connection loss exceptions are
ignored, and all other exceptions are logged as warnings.

Fixes: MI-1276

Change-Id: If18346ef6ba65394b6f01b3cb1de605a3923cff1
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>