FlowState: reuse receiving message buffers
We were allocating a 1KB buffer for every new message processed by
the flow state minion which translated to a lot of garbage. While
the GC pauses on the flow state minion are not the bottleneck to
process messages, they can be certainly reduced by caching the
message buffer locally to each processing thread.

This patch also changes how the message is read from the underlying
direct buffer provided by Netty. Before this patch, two new
ByteBuffers (message header and message body) weere allocated when
calling the nioBuffer method. This patch caches these byte buffers
locally in a thread and copies the content from the underlying buffer
directly. This optimization also highly reduces the amount of garbage
generated and reduces the amount of GCs.

Change-Id: I55ffbba5a4e55661c41a5e59999fd84b9c758836
Signed-off-by: Xavi León <xavi.leon@midokura.com>