FlowState: reuse receiving message buffer
We were allocating a 1KB buffer for every new message processed by
the flow state minion which translated to a lot of garbage. While
the GC pauses on the flow state minion are not the bottleneck to
process messages, they can be certainly reduced by caching the
message buffer locally to each processing thread.

Change-Id: I55ffbba5a4e55661c41a5e59999fd84b9c758836
Signed-off-by: Xavi León <xavi.leon@midokura.com>