Remove infinite loop in ArpRequestBroker
ArpRequestBroker#shouldProcess uses ||= in a loop with an iterator. ||=
evaluates lazily, which means the right hand side is only evaluated if
the left hand side is false. Since the call to next() was part of the
RHS, if tables.next().shouldProcess() ever returned true, next()
wouldn't ever be called again and hasNext() would always evaluate to
true, giving us an infinite loop. The issue has been fixed.

Change-Id: I74fb825fc110d867d360ecad7e405dd537dfd322
Signed-off-by: Ivan Kelly <ivan@midokura.com>