Allow blocks to be memory mapped at higher offsets
We were using an Int to specify the offset of the memory mapped block,
which limited us to a max of 2 Gb offsets. Move to use a Long so we
don't hit that limit in the foreseeable future.

Change-Id: I5dab0dc7b946a43148c97266c2054f340714d28a
Signed-off-by: Xavi León <xavi.leon@midokura.com>
(cherry picked from commit 94b3cd6a95763bb7fa03098b0b645dfd65a39ef3)