10kp: State table cache drops old updates
Currently the state table cache processes all updates notified
from ZooKeeper on a single threaded executor. When a notification
occurs, a processing task is scheduled on the executor queue,
which is responsible to compute the changes to the state table
entries, compute the update notifications and send them to the
corresponding client subscribers.

However, when the ZooKeeper notification rate is higher than the
rate at which the cache can process the changes and send the
messages to the subscribers, processing tasks will accummulate
on the executor queue delaying both the processing of the latest
changes and sending of updates to subscribers.

This patch fixes the execution queueing delay issue by introducing
a size-one queue that will store only the latest notification from
ZooKeeper, dropping all previous notifications that were not yet
processed. This ensures that when the cache cannot keep up with
processing all changes, it will prioritize the latest, and discarding
older ones.

Fixes: MI-1298

Change-Id: I838c2815141c87c3ee783407410f75759b186b81
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit dd9ca910305f412922e4916c904d283f9aa2734a)