FlowState: close file descriptors when removing a port
The flow state minion uses a file for each active port bound. When
releasing a port, we were not closing the file channel so we didn't
close the associated file descriptor. This patch makes sure we
close the file channel when removing the port and cleans up a few
exception messages (so we log and throw as soon as possible) so the
users catching the exception doesn't need to log it again.

Ref.: MI-1297
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: I6133d282d2e5f44ae4e90b682ce87d80b73713a1