FlowState: reset block when removing from the ring buffer
When we detect that a ring buffer is full and we need to write to it, we
remove the tail and allocate a new block with the corresponding factory.
When using a byte buffers backed by heap memory there is no issue as the
factory will detect that the byte buffer is empty/invalid and initialize
it properly.

However, when using byte buffers backed by memory mapped files and
reusing the same memory regions, the factory detected that the block was
valid and set its position to the end of the block. This made subsequent
writes after wrapping around the ring buffer to fail (because we were
trying to write past the block limit).

This patch makes sure that when wrapping around the ring buffer and
overwriting valid blocks, we first reset it to mark it as invalid. Just
like when we run the block invalidator.

Ref.: MI-1289
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: Idfc2e00dc7524e88f1ac02ca91d4a5b8de75036c
(cherry picked from commit 195e7c374ddc4314b99be2f95706a67094b44c34)