Do not reuse context on reader handler
We were reusing the connection context among several incoming connection
to the flow state minion. This could make answers to a given request go
into another request. This patch just passes the context around as a
parameter to all functions so each connection is associated with its
context.

Ref.: MI-1299
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: Ic8d11ccd495427f130d14ba7d131687d418b079f