FlowState: do not load message blocks unless necessary
We only need to actually read, decode and allocate new objects
when writing to legacy storage, as the Cassandra clients expects
decoded objects. However, when writing into local storage, we write the
SBE directly without decoding most of the message (only the last block
with the ingress and egress port UUIDs are necessary).

This patch treats the message differently taking into account if we
we need to store it to cassandra or write it to local storage.

Change-Id: I71ec1a9ac2fc4fd9761da51f083d842e1ade2f22
Signed-off-by: Xavi León <xavi.leon@midokura.com>