FlowState: sanitize configuration defaults
We were setting a very large file for all kinds of nodes. However,
requirements for a gateway and a compute are different. This patch
changes the default values considering a compute node (lots of virtual
ports/vms that can fit on small files). It also updates the templates
for the gateway nodes to increase the file size (number of blocks and
size of the block) as it will receive flow state from the other gateways
as well as from all computes. Look at the configuration option
descriptions for a more comprehensive explanation.

Change-Id: I799d3e55b0ce5b3e6ab8b8b27390f9f7cb2bfdcd
Signed-off-by: Xavi León <xavi.leon@midokura.com>