Recycler: Fix flakey RecyclerTest
A race-condition has been found in five scenarios of RecyclerTest

If an object is created and the recycler is run within the same millisecond
the recycler will not remove the object as it's expecting it to have been
created strictly in the past.

This fix makes sure that sufficient time passes before running the recycler
task, in those tests expecting objects to be removed.

Change-Id: If9a8359dc2b1699c6b2d7bf582c13dc17088be71
Ref: MI-1312