API: change read timeout config
Read timeouts were not being read from config but were hard-coded to a
rather large value of 30 seconds. This caused the client requests to
hang until they timeout from their side or the grace period is over and
the agent dies. For large grace times this could cause long waits on the
client side of the API.

The timeout is now read from configuration and by default lowered to 2
seconds.

Ref: MI-436

Change-Id: I3a2c3d2003d6ecf462ad83731fc08006e8051829
Signed-off-by: Mateo Hermida <mateo@midokura.com>