Midonet CLI: incorrect requests made for port
When listing port descriptions, the CLI always made calls as if the port
was a Bridge AND a Router port. One of these failed and caused a 404.
This was harmless though, as the exception was catched and the next
request was made. However, it still adds an unnecessary request, and
adds noise to the logs.

Ref: MI-734

Change-Id: I492c79dbe1204dc7f1a606aedad2a7dd0644ca8f
Signed-off-by: Mateo Hermida <mateo@midokura.com>