Metrics: Expose higher level storage errors
Currently we are exposing the lower level storage errors such as
KeeperExceptions errors. The purpose of the ZOOM storage is to hide
the ZooKeeper node manipulation with a class/object paradigm.
However, the metrics do not expose the errors at this higher level.
In addition, we record the error metrics both at the level of the
internal components such as NodeObservable, DirectoryObservable, as
well as the external components that use them such as the
ZookeeperObjectMapper and ZookeeperObjectState classes, which
may lead to recording some exceptions several times.

This patch changes the way we exposes storage error metrics,
replacing them with only higher level errors, corresponding to
exceptions raised during ZOOM object manipulations.

The recorded error metrics are the following:

* Metrics on ZOOM writes errors: conflict, concurrent modification,
  node exists, node not found, object exists, object not found, or
  other internal error.

* Metrics on ZOOM reads: object not found, or internal errors.

* Metrics on ZOOM observables: recovery on closed observables, and
  observable errors.

Change-Id: I9a9aba6d770af71f02f9f5d4b94a21e272162f3e
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>