10kp: Improve performance of table cache
This patch includes a couple of improvements that increase the
StateTableCache performance when dealing with a large number of
entries.

- Use strings as hashmap keys instead of protocol buffers classes

It was observed that calculating the hashcode of such class and
comparing them was a significant burden to execution times.

- Reserve enough capacity to hold all map entries

This prevents the cache HashMap from being resized multiple times
when it's populated, as we know the exact capacity required.

Change-Id: I5c8df3895098c2637a57279a58520d174685a2eb