Updating ARP entries related to FIPs on GW port MAC change
In case the MAC is change in a port that is acting as gateway for
FIPs, the entries in the associated network ARP tables has to be
updated, deleting entry of old MAC and creating one for new MAC.

This is done in this commit by implementing a new method in the
PortTranslator that find impacted FIPs by checking the NAT rules of
the port in case there is a MAC update.

A unit test is also included, that check the create/delete
operations are added by the PortTranslator.

Ref: MI-1210

Change-Id: Ie1bbf56e36caf11092745b22fdda5422c60944ed
(cherry picked from commit 5e64ca4de461a3b882b9b3815bbb3209e8e5426e)