Add subnet to NeutronNetwork subnet list
The Neutron models are not updated when associated objects are created.
In this case, the NeutronNetwork has a list of subnets which is always
empty because neutron does not send over an updated network when we
create a new network.

This patch gets around this by modifying the neutron model explicitly
when a subnet is created or deleted.

Change-Id: Iba64b9b50feaa6f6ef3ea021688797c12a58fcf4