Fix: Expire flows after inactivity
Flows were not being expired if no packets are being received in the
agent. This means that if no packets were being received, then the next
incoming packet would still see an invalid stale flow that should have
been expired.

To fix this we add a check on the timeout call of the disruptor to see
when was the table expired for the last time. If it was over 5 seconds
then we call the processing (and expiring) function again.

Change-Id: I7e668c6e8c497f17e4fe211d89d198493567c896
Signed-off-by: Mateo Hermida <mateo@midokura.com>