Add route for all subnets to router
Floating IPs can come from any subnet associated with an external
network. However, when hooking up a router to an external network,
we only add a route for the first subnet. This means that routers
have no way to route or arp for floating IPs allocated from
second and third subnets.

This patch addresses this problem by adding a second route on
all routers connected to a subnet if:
   1) a second (or third or fourth) subnet is added to a network.
   2) a router interface is created on a subnet whose network
      has multiple subnets.

Fixes MN-953

Change-Id: I71e1955b441a3e0576ed3a9ec239bb497a13f1e4