Add fip_nat_rule_ids to the REST API Port
This is necessary to ensure updates to the port via the REST API
still keep this data if it's set.

Marked this as JSONIgnore so it doesn't get exposed via the API,
as this is a private data field.

Change-Id: I5a6278c2b3a20a6c6334a42e2dc3579a8fe19ae5
Fixes-bug: MI-1355
Signed-off-by: Michael Micucci <micucci@midokura.com>
(cherry picked from commit f8f38c11850d3c07604fb8a48fe2367bf7509479)