Recycler: Fix recycler fails to delete state
This fixes an issue where the recycler cannot delete a state key
for an object, because the node has children. This happens for
multi-value keys, where the values are stored as children for the
state key node rather than data in the key node.

The patch modifies the deletion logic to use the `deleteWithChildren`
method rather than a regular delete.

Fixes: MI-1345

Change-Id: I83a91b68c508e25e96ebcfaba55b104a68579f8b
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>