Update patch set 1
Patch Set 1:

(4 comments)

For the value, I think 15-30 seconds is reasonable to accommodate the latency for most requests. Btw, do we know why there is a difference between 1.9 and 5.x with respect to the timeout? In 1.9 the read from ZK in synchronous, there is no timeout at all. Why does that fail sooner than reading from ZK in 5.x?

Patch-set: 1
Label: Verified=0