BGP: Use MidoNet logger
This changes the logger in from the BGP classes (routing manager
actor, routing handler, BGP daemon and VTY connection) to use
the MidoNet logger.

This avoids the creation of a new logger, which will eventually
leak, for each new BGP routing instance, and adds context
information to the BGP daemon and VTY connection.

Change-Id: Ic0f6cbec1831e1d2cbfbd4a1277d18f969bab6c9
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>