Update patch set 2
Patch Set 2:

> (4 comments)
 > 
 > For the value, I think 15-30 seconds is reasonable to accommodate
 > the latency for most requests. Btw, do we know why there is a
 > difference between 1.9 and 5.x with respect to the timeout? In 1.9
 > the read from ZK in synchronous, there is no timeout at all. Why
 > does that fail sooner than reading from ZK in 5.x?

I changed the comments, and left the timeout in 30s. I haven't tried it, I'm guessing it checks the status of the connection or something. We could do a check to see if the session is not connected and fail fast in that case maybe, what do you think?

Patch-set: 2
Reviewer: Mateo Hermida <1008328@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0